clean up `fs` and `platform` extensions
